+++ CHANGELOG.md
@@ -3,6 +3,8 @@ Changes to the schema are documented in this file.
 
 ## Current (updates before releasing)
 ### Added
+  - `relevant protocols` and `funding sources` slots
+  - `ess dive datasets` slot
   - add amplicon slots to omics processing (#139)
 ### Fixed
   - N/A